fix three issues that strict clang on macOS would flag
1. src/arm/cc1200.c: stray "/*" inside a doc comment (left over from
   a prior comment-block edit). gcc emits -Wcomment; clang -Werror in
   stricter configs would reject it.

2. src/arm/nrf52840_soc.c: memcpy() was used at line 595 without
   including <string.h>. Implicit-function-declaration is an error
   under strict C11 (clang on macOS defaults closer to that than gcc).

3. test/test_arm_correctness.c: SMLATT overflow test summed two
   signed-int literals (0x7FFFFFFF + 0x3FFF0001) which is UB in C
   even though the test is *checking* the wraparound result. Cast
   each operand to uint32_t first so the wrap happens in unsigned
   arithmetic (well-defined) and the cast(uint32_t)(...) result is
   the same.

These are the clang-strict-mode candidates for the macOS CI failure
on PR #4. If the actual failure is a different step entirely (e.g. a
runtime test diverging on macOS), share the exact error and I'll
target that specifically.

+++ test/test_arm_correctness.c
@@ -669,7 +669,7 @@ static void test_m4_dsp_halfword_multiply(void) {
         arm_step(&cpu, 1);
         /* Sum = 0x7FFFFFFF + 0x3FFF0001 overflows int32 → Q sticky */
         assert_eq("SMLATT R2 = wrapped sum",
-                  (uint32_t)(0x7FFFFFFF + 0x3FFF0001), cpu.reg[2]);
+                  (uint32_t)0x7FFFFFFF + (uint32_t)0x3FFF0001, cpu.reg[2]);
         assert_true("SMLATT overflow sets Q", (cpu.xpsr & APSR_Q) != 0);
         arm_cpu_destroy(&cpu);
     }
